+++ libcaf_io/src/io/basp_broker.cpp
@@ -366,8 +366,7 @@ behavior basp_broker::make_behavior() {
         return unit;
       return sec::cannot_close_invalid_port;
     },
-    [=](get_atom,
-        const node_id& x) -> std::tuple<node_id, std::string, uint16_t> {
+    [=](get_atom, const node_id& x) -> result<node_id, std::string, uint16_t> {
       std::string addr;
       uint16_t port = 0;
       auto hdl = instance.tbl().lookup_direct(x);
@@ -375,7 +374,7 @@ behavior basp_broker::make_behavior() {
         addr = remote_addr(*hdl);
         port = remote_port(*hdl);
       }
-      return std::make_tuple(x, std::move(addr), port);
+      return {x, std::move(addr), port};
     },
     [=](tick_atom, size_t interval) {
       instance.handle_heartbeat(context());